Fix chain_walker to get current depth of chain

+++ src/config/genesis.rs
@@ -19,7 +19,7 @@ pub struct Genesis {
 /// A configuration element for clique specific values.
 #[derive(Serialize, Deserialize, Debug)]
 pub struct CliqueConfig {
-    pub block_period: usize,
+    pub block_period: u64,
     pub signer_limit: usize,
 }
